+++ mediavocab/helpers/__init__.py
@@ -7,10 +7,12 @@ from mediavocab.helpers.classify import (
 )
 from mediavocab.helpers.queries import (
     credits_with_role, primary_credit, director, author, performers,
+    episodes_of, filmography_of, quality_score, best_release,
 )
 
 __all__ = [
     "make_movie", "make_episode", "make_release", "make_credit",
     "is_not_media", "is_generic", "is_device_entity", "is_continuous_release",
     "credits_with_role", "primary_credit", "director", "author", "performers",
+    "episodes_of", "filmography_of", "quality_score", "best_release",
 ]