+++ mediavocab/taxonomy/__init__.py
@@ -7,6 +7,7 @@ from mediavocab.taxonomy.relation import (
     RelationRole,
     CreditSection,
     WorkRelationKind,
+    ReleaseRelationKind,
 )
 from mediavocab.taxonomy.membership import MembershipStatus
 from mediavocab.taxonomy.content_type import ContentType
@@ -40,8 +41,8 @@ from mediavocab.taxonomy.genre import (  # noqa: F401  (re-exported)
 
 __all__ = [
     "MediaType", "VariantKind", "ReleaseStatus", "StreamMode", "EntityKind",
-    "RelationRole", "CreditSection", "WorkRelationKind", "MembershipStatus",
-    "ContentType",
+    "RelationRole", "CreditSection", "WorkRelationKind", "ReleaseRelationKind",
+    "MembershipStatus", "ContentType",
     # genres
     "GENRE_DOCUMENTARY", "GENRE_ANIMATION", "GENRE_ANIME", "GENRE_SHORT_FILM",
     "GENRE_NOIR", "GENRE_CONCERT", "GENRE_STAND_UP", "GENRE_TALK_SHOW",